refactor(): use hyphens for filenames
