+++ src/core/index.ts
@@ -1,1 +1,1 @@
-export * from './public_api';
+export * from './public-api';
