+++ src/core/shared/logger/logger.ts
@@ -1,6 +1,6 @@
 import { transports, Logger as winstonLogger } from 'winston';
 
-import { isDevMode } from '../../orbital_ref';
+import { isDevMode } from '../../orbital-ref';
 import { format } from './format';
 
 /**
